Labs/GCP_Labs/CloudFunction_Labs/Lab2 - CloudFunction_Intermediate/src/serving/requirements.txt
google-cloud-storage==2.7.0
joblib==1.2.0
scikit-learn>=1.0
numpy>=1.21
